requirements.txt
Django==4.2.11
djangorestframework==3.14.0
python-decouple==3.8
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.31.0
celery==5.3.4
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
# req for geo data processing. 
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5
numpy===1.24.3
